Split InstallStep execution into multiple classes
