+++ cloudnet/src/main/java/de/dytanic/cloudnet/template/install/run/step/executor/CopyFilterStepExecutor.java
@@ -1,0 +1,57 @@
+package de.dytanic.cloudnet.template.install.run.step.executor;
+
+
+import com.google.gson.reflect.TypeToken;
+import de.dytanic.cloudnet.template.install.run.InstallInformation;
+import org.jetbrains.annotations.NotNull;
+
+import java.io.IOException;
+import java.lang.reflect.Type;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.StandardCopyOption;
+import java.util.*;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
+
+public class CopyFilterStepExecutor implements InstallStepExecutor {
+
+    private static final Type STRING_MAP_TYPE = new TypeToken<Map<String, String>>() {
+    }.getType();
+
+    @Override
+    public @NotNull Set<Path> execute(@NotNull InstallInformation installInformation, @NotNull Path workingDirectory, @NotNull Set<Path> inputPaths) throws IOException {
+        Map<String, String> copy = installInformation.getServiceVersion().getProperties().get("copy", STRING_MAP_TYPE);
+
+        if (copy == null) {
+            throw new IllegalStateException(String.format("Missing copy property on service version %s!", installInformation.getServiceVersion().getName()));
+        }
+
+        List<Map.Entry<Pattern, String>> patterns = copy.entrySet().stream()
+                .map(entry -> new AbstractMap.SimpleEntry<>(Pattern.compile(entry.getKey()), entry.getValue()))
+                .collect(Collectors.toList());
+
+        Set<Path> resultPaths = new HashSet<>();
+
+        for (Path path : inputPaths) {
+            String relativePath = workingDirectory.relativize(path).toString().replace("\\", "/").toLowerCase();
+
+            for (Map.Entry<Pattern, String> patternEntry : patterns) {
+                Pattern pattern = patternEntry.getKey();
+                Path targetPath = workingDirectory.resolve(patternEntry.getValue().replace("%fileName%", path.getFileName().toString()));
+
+                if (pattern.matcher(relativePath).matches()) {
+                    Files.copy(path, targetPath, StandardCopyOption.REPLACE_EXISTING);
+                    resultPaths.add(targetPath);
+                }
+            }
+        }
+
+        return resultPaths;
+    }
+
+    @Override
+    public void interrupt() {
+    }
+
+}
